# fixed Conflicts:
#	src/Engine/Core.cs
#	src/Engine/Game_Engine.cs
#	src/Engine/Subsystems/Position_Integrator.cs
#	src/Engine/Subsystems/Sprite_Renderer.cs
#	src/Game.cs
#	src/Game/Entities/Ball.cs

+++ src/Engine/Components/FpsCounter.cs
@@ -6,7 +6,7 @@ namespace Engine.Components
 {
     public class FpsCounter : Base_Component
     {
-        public int frameRate { get; set; }
-        public int frameCounter { get; set;}
+        public float frameRate { get; set; }
+        public float frameCounter { get; set;}
     }
 }
